+++ abserde-cli/src/commands/mod.rs
@@ -2,4 +2,5 @@ pub mod auth;
 pub mod deauth;
 pub mod init;
 pub mod new;
+pub mod rtbf;
 pub mod update;